Refactored and added list_items function and get_input functions
Uses list_items function to echo items to the screen and hide index 0
each time it is called. Uses get_input to manipulate the user input –
if it should be uppercase it changes it to uppercase. Removed “ghost”
index and used $key = key + 1 and $key = key -1 to make list count
start at 1.